Add -repl flag and set default behaviour to exit on error
This changes the default `error-behaviour` to `immediate-exit` for smtlib2 when running in CLI mode. This makes it fail-safe. There's no reason to ignore errors in CLI mode and it may lead to accidentally "verified" proofs if the user makes an easy mistake.

This does not change the default when used via an API because `immediate-exit` calls `_Exit(1)` and that isn't a good thing for a library to do.

When passing commands via standard input there are two use cases:

* Running Z3 from another program (e.g. a compiler), in which case you don't want to ignore errors.
* Using it as a basic REPL, in which case you don't.

I added a `-repl` flag for the second use case. This also opens the door for potentially adding a "proper" REPL in future (with history, etc.).